iternet/requirements.txt
torch
torchvision
fastai==1.0.60
LMDB
Pillow
opencv-python
tensorboardX
editdistance
synthtiger